App/python-server/requirements.txt
fastapi
asyncpg
passlib[bcrypt]
python-jose
uvicorn
pydantic[email]
dotenv
httpx
asyncio
apscheduler
requests
gradio
fastapi-mail
